passing instantiation tests

+++ tests/test_hitandrun.py
@@ -15,9 +15,9 @@ class TestHitAndRun(unittest.TestCase):
         A = np.array([[1, 0],
                       [-1, 0],
                       [0, 1],
-                      [0, -1]])
-        b = np.array([1, 1, 1, 1])
-        x0 = np.array([-.5, -.5])
+                      [0, -1]], dtype=np.float32)
+        b = np.array([1, 1, 1, 1], dtype=np.float32)
+        x0 = np.array([-.5, -.5], dtype=np.float32)
         polytope = Polytope(A=A, b=b)
         hitandrun = HitAndRun(polytope=polytope, starting_point=x0)
         self.assertTrue(isinstance(hitandrun, HitAndRun))
