+++ tests/test_polytope.py
@@ -10,8 +10,8 @@ class TestPolytope(unittest.TestCase):
         A = np.array([[1, 0],
                       [-1, 0],
                       [0, 1],
-                      [0, -1]])
-        b = np.array([1, -1, 1, -1])
+                      [0, -1]], dtype=np.float32)
+        b = np.array([1, -1, 1, -1], dtype=np.float32)
         polytope = Polytope(A=A, b=b)
 
         self.assertTrue(isinstance(polytope, Polytope))
@@ -20,8 +20,8 @@ class TestPolytope(unittest.TestCase):
         A = np.array([[1, 0],
                       [-1, 0],
                       [0, 1],
-                      [0, -1]])
-        b = np.array([1, -1, 1, -1])
+                      [0, -1]], dtype=np.float32)
+        b = np.array([1, -1, 1, -1], dtype=np.float32)
         polytope = Polytope(A=A, b=b)
 
         self.assertEqual(polytope.dim, 2)
